feat(githubaction): adapt for use as a github action
Uses the existing docker container

The 'scan-url' Github workflow argument is explicity mapped to the
SCAN_URL environment variable already expected&handled by the docker
container.

This enables the Dockerfile to remain agnostic to the fact that it is
running as a Github action, rather than depending on the 'INPUT_'
github action var naming.